requirements.txt
sqlalchemy
psycopg2-binary
matplotlib
seaborn
